plataforma/backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.2
psycopg[binary]>=3.2.10
dj-database-url==2.2.0
python-decouple==3.8
openpyxl==3.1.5
python-docx==1.1.2
drf-spectacular==0.27.2
